--- achievements.go
@@ -1,27 +1,0 @@
-package gopixel
-
-import (
-	"encoding/json"
-
-	"errors"
-
-	"github.com/comblock/gopixel/structs"
-)
-
-func (client *Client) Achievements() (structs.Achievements, error) {
-	var achievements structs.Achievements
-
-	data, err := get("api.hypixel.net/resources/achievements?key=" + client.Key)
-
-	if err != nil {
-		return achievements, err
-	}
-
-	err = json.Unmarshal(data, &achievements)
-
-	if !achievements.Success {
-		err = errors.New(achievements.Cause)
-	}
-
-	return achievements, err
-}
