--- boosters.go
@@ -1,20 +1,0 @@
-package gopixel
-
-import (
-	"encoding/json"
-
-	structs "github.com/comblock/gopixel/structs"
-)
-
-func (client *Client) Boosters() (structs.Boosters, error) {
-	var boosters structs.Boosters
-
-	data, err := get("api.hypixel.net/boosters?key=" + client.Key)
-	if err != nil {
-		return boosters, err
-	}
-
-	err = json.Unmarshal(data, &boosters)
-
-	return boosters, err
-}
