--- friends.go
@@ -1,32 +1,0 @@
-package gopixel
-
-import (
-	"encoding/json"
-
-	"errors"
-
-	structs "github.com/comblock/gopixel/structs"
-)
-
-func (client *Client) Friends(name string) (structs.Friends, error) {
-	var friends structs.Friends
-
-	uuid, err := Uuid(name)
-
-	if err != nil {
-		return friends, err
-	}
-	data, err := get("api.hypixel.net/friends?uuid=" + uuid + "&key=" + client.Key)
-
-	if err != nil {
-		return friends, err
-	}
-
-	err = json.Unmarshal(data, &friends)
-
-	if !friends.Success {
-		err = errors.New(friends.Cause)
-	}
-
-	return friends, err
-}
