+++ guild.go
@@ -5,8 +5,34 @@ import (
 
 	"errors"
 
-	structs "github.com/comblock/gopixel/structs"
+	"github.com/comblock/gopixel/structs"
 )
+
+func (c *Client) GuildAchievements() (structs.GuildAchievements, error) {
+	var guildAchievements structs.GuildAchievements
+
+	data, err := get("api.hypixel.net/resources/guilds/achievements?key=" + c.Key)
+	if err != nil {
+		return guildAchievements, err
+	}
+
+	err = json.Unmarshal(data, &guildAchievements)
+
+	return guildAchievements, err
+}
+
+func (client *Client) GuildPermissions() (structs.GuildPermissions, error) {
+	var guildPermissions structs.GuildPermissions
+
+	data, err := get("api.hypixel.net/resources/guilds/permissions?key=" + client.Key)
+	if err != nil {
+		return guildPermissions, err
+	}
+
+	err = json.Unmarshal(data, &guildPermissions)
+
+	return guildPermissions, err
+}
 
 func (client *Client) GuildById(id string) (structs.Guild, error) {
 	var guild structs.Guild
