--- leaderboards.go
@@ -1,20 +1,0 @@
-package gopixel
-
-import (
-	"encoding/json"
-
-	structs "github.com/comblock/gopixel/structs"
-)
-
-func (client *Client) Leaderboards() (structs.Leaderboards, error) {
-	var leaderboards structs.Leaderboards
-
-	data, err := get("api.hypixel.net/leaderboards?key=" + client.Key)
-	if err != nil {
-		return leaderboards, err
-	}
-
-	err = json.Unmarshal(data, &leaderboards)
-
-	return leaderboards, err
-}
