--- player_counts.go
@@ -1,19 +1,0 @@
-package gopixel
-
-import (
-	"encoding/json"
-
-	structs "github.com/comblock/gopixel/structs"
-)
-
-func (client *Client) PlayerCounts() (structs.PlayerCounts, error) {
-	var counts structs.PlayerCounts
-	data, err := get("api.hypixel.net/counts?key=" + client.Key)
-	if err != nil {
-		return counts, err
-	}
-
-	err = json.Unmarshal(data, &counts)
-
-	return counts, err
-}
