--- quests.go
@@ -1,20 +1,0 @@
-package gopixel
-
-import (
-	"encoding/json"
-
-	structs "github.com/comblock/gopixel/structs"
-)
-
-func (client *Client) Quests() (structs.Quests, error) {
-	var quests structs.Quests
-
-	data, err := get("api.hypixel.net/resources/quests?key=" + client.Key)
-	if err != nil {
-		return quests, err
-	}
-
-	err = json.Unmarshal(data, &quests)
-
-	return quests, err
-}
