--- skyblock_skills.go
@@ -1,20 +1,0 @@
-package gopixel
-
-import (
-	"encoding/json"
-
-	structs "github.com/comblock/gopixel/structs"
-)
-
-func (client *Client) SkyblockSkills() (structs.SkyblockSkills, error) {
-	var skyblockSkills structs.SkyblockSkills
-
-	data, err := get("api.hypixel.net/resources/skyblock/skills?key=" + client.Key)
-	if err != nil {
-		return skyblockSkills, err
-	}
-
-	err = json.Unmarshal(data, &skyblockSkills)
-
-	return skyblockSkills, err
-}
